@@ -19,7 +19,7 @@ import { WegvakonderdeelplanningControllerService } from './api/wegvakonderdeelp
     WegvakonderdeelplanningControllerService ]
 })
 export class ApiModule {
-    public static forRoot(configurationFactory: () => Configuration): ModuleWithProviders {
+    public static forRoot(configurationFactory: () => Configuration): ModuleWithProviders<ApiModule> {
         return {
             ngModule: ApiModule,
             providers: [ { provide: Configuration, useFactory: configurationFactory } ]
